Refactor into classes.
